Derive Macro amethyst_derive::SystemDesc

#[derive(SystemDesc)]
{
    // Attributes available to this derive:
    #[system_desc]
}

Derive a SystemDesc implementation.

The SystemDesc is passed to the GameData to instantiate the System when building the dispatcher.

This derive may be used for Systems that do not require special code for System::setup.